api-connector/requirements.txt
pyserial==3.5
requests==2.26.0
python-dotenv==0.19.1
tweepy